feat(scripts): generate buildspec.yml and circle.yml after generate-included-image script is run
